address cr comments
